made backlink compact + removed comments + moved filelist bundle to group
