+++ lib/themes/eic_community/styleguide/bundles/group.stories.js
@@ -10,6 +10,8 @@ import DiscussionFullTemplate from '@theme/patterns/compositions/group/group.ful
 import DiscussionThreadTemplate from '@theme/patterns/compositions/discussion-thread.html.twig';
 import EventTeaserTemplate from '@theme/patterns/compositions/event/event.teaser.html.twig';
 import ExtendedListTemplate from '@theme/patterns/compositions/extended-list/extended-list.html.twig';
+import FileListFullTemplate from '@theme/patterns/compositions/filelist/filelist.full.html.twig';
+import FilelistTemplate from '@theme/patterns/components/filelist.html.twig';
 import MemberFullTemplate from '@theme/patterns/compositions/group/group.full--member.html.twig';
 import MemberTemplate from '@theme/patterns/compositions/member/member.teaser.html.twig';
 import FrontpageTemplate from '@theme/patterns/compositions/group/group.full--frontpage.html.twig';
@@ -31,20 +33,25 @@ import author from '@theme/data/author.data';
 import breadcrumb from '@theme/data/breadcrumb.data';
 import common from '@theme/data/common.data';
 import comments from '@theme/data/comments.data';
+import contributors from '@theme/data/contributors.data';
 import discussionThread from '@theme/data/discussion-thread.data';
 import extendedList from '@theme/data/extended-list';
 import featuredContentCollection from '@theme/data/featured-content-collection';
 import featuredContentSections from '@theme/data/featured-content-sections';
+import filelist from '@theme/data/filelist.data';
 import filters from '@theme/data/filters';
+import hero from '@theme/data/hero.data';
 import mainmenu from '@theme/data/mainmenu.data';
 import pagination from '@theme/data/pagination.data';
 import searchform from '@theme/data/searchform.data';
 import siteFooter from '@theme/data/site-footer.data';
 import siteHeader from '@theme/data/site-header.data';
+import socialShare from '@theme/data/social-share.data';
 import sorting from '@theme/data/sorting.data';
 import subnavigation from '@theme/data/subnavigation';
 import teaser from '@theme/data/teaser';
 import teaserOverview from '@theme/data/teaser-overview.data';
+import topics from '@theme/data/topics.data';
 import topMenu from '@theme/data/top-menu.data';
 import wiki from '@theme/data/wiki';
 
@@ -1103,3 +1110,94 @@ export const OverviewPrivate = () =>
     ),
     pagination: pagination,
   });
+
+
+
+  
+ 
+  
+  
+  let filelistDetail = {
+    common: common,
+    site_footer: siteFooter,
+    site_header: siteHeader,
+    searchform: searchform,
+    editorial_header: {
+      icon_file_path: common.icon_file_path,
+      parent: {
+        link: {
+          label: 'Previous page',
+          path: '#',
+        },
+      },
+    },
+    hero: hero,
+    breadcrumb: breadcrumb,
+    editorial_actions: {
+      icon_file_path: common.icon_file_path,
+      items: [
+        {
+          link: {
+            label: 'Liked (27)',
+            path: '?path=like',
+          },
+          icon: {
+            type: 'custom',
+            name: 'like',
+          },
+        }
+      ],
+    },
+    contributors: contributors,
+    social_share: socialShare,
+    topics: topics,
+    content: FilelistTemplate(filelist),
+  };
+  
+  export const FileListFullPublic = () => FileListFullTemplate(filelistDetail);
+  
+  export const FileListFullPrivate = () => FileListFullTemplate(
+    Object.assign(
+      {},
+      filelistDetail,
+      {
+        user: common.user,
+        editorial_actions: {
+          icon_file_path: common.icon_file_path,
+          items: [
+            {
+              link: {
+                label: 'Bookmarked',
+                path: '?path=bookmark',
+              },
+              icon: {
+                type: 'custom',
+                name: 'tag',
+              },
+            },
+            {
+              link: {
+                label: 'Liked (27)',
+                path: '?path=like',
+              },
+              icon: {
+                type: 'custom',
+                name: 'like',
+              },
+            },
+            {
+              link: {
+                label: 'Add to highlight',
+                path: '?path=login',
+              },
+              icon: {
+                type: 'custom',
+                name: 'star_circle',
+              },
+            }
+          ],
+        },
+      }
+    )
+  )
+
